+++ inject/inject-utils/src/test/scala/com/twitter/inject/tests/conversions/StringConversionsTest.scala
@@ -5,14 +5,6 @@ import com.twitter.inject.conversions.string._
 
 class StringConversionsTest extends Test {
 
-  test("toOption#when nonEmpty") {
-    "foo".toOption should be(Some("foo"))
-  }
-
-  test("toOption#when empty") {
-    "".toOption should be(None)
-  }
-
   test("ellipse") {
     "foobar".ellipse(2) should be("fo...")
   }